+++ agent/src/main/java/com/walmartlabs/concord/agent/logging/RedirectedProcessLog.java
@@ -58,6 +58,10 @@ public class RedirectedProcessLog implements ProcessLog {
         streamLog(localLog.logFile(), stopCondition, logSteamMaxDelay, sink);
     }
 
+    public void flush() throws Exception {
+        run(() -> true);
+    }
+
     @Override
     public void delete() {
         this.localLog.delete();